Take a node reference on the vap bss node before doing anything.
Note: this doesn't fix the race condition; the problem with vap bss
node handling is that it can be ripped out in parallel whilst a state
change is occuring.  The complete lack of sensible locking in iwm_newstate()
is also a problem.  This is just me attempting to make it less likely.
Sigh.